% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_dia}
\alias{gh_dia}
\alias{gh_ppt}
\alias{gh_hamta_analytikernatverket}
\title{Snabblistning av filer i vanliga repos}
\usage{
gh_dia(filter = NULL)

gh_ppt(filter = NULL)

gh_hamta_analytikernatverket(filter = NULL)
}
\arguments{
\item{filter}{Sök-DSL, se \code{\link[=github_lista_repo_filer]{github_lista_repo_filer()}}.}
}
\value{
Se \code{\link[=github_lista_repo_filer]{github_lista_repo_filer()}}.
}
\description{
\code{gh_dia()} listar \code{diagram}-repot, \code{gh_ppt()} samma som \code{ppt}-kodrader,
\code{gh_hamta_analytikernatverket()} listar Analytikernätverkets \code{hamta_data}.
}
