% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_commit_push}
\alias{github_commit_push}
\alias{github_commit_push_analytikernatverket}
\title{Committa och pusha alla ändringar i ett lokalt repo}
\usage{
github_commit_push(
  repo,
  commit_txt = NULL,
  grundsokvag = intern_gh_mapp(),
  pull_forst = TRUE
)

github_commit_push_analytikernatverket(
  repo,
  commit_txt = NULL,
  grundsokvag = intern_gh_mapp_an(),
  pull_forst = TRUE
)
}
\arguments{
\item{repo}{Reponamn.}

\item{commit_txt}{Commit-meddelande. \code{NULL} = genereras ur git-status.}

\item{grundsokvag}{Föräldermapp. Standard: \link{rddeploy-config} \code{gh_mapp}.}

\item{pull_forst}{Kör \verb{git pull} innan push (om det finns en upstream-branch).}
}
\value{
Osynligt \code{TRUE} om något pushades, annars \code{FALSE}.
}
\description{
Committa och pusha alla ändringar i ett lokalt repo
}
