% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_lagg_till_repo_fran_github}
\alias{github_lagg_till_repo_fran_github}
\alias{github_lagg_till_repo_fran_github_analytikernatverket}
\title{Klona ett GitHub-repo lokalt}
\usage{
github_lagg_till_repo_fran_github(
  repo_namn,
  repo_org = intern_gh_org(),
  grundsokvag = intern_gh_mapp(),
  oppna_rproj = FALSE
)

github_lagg_till_repo_fran_github_analytikernatverket(
  repo_namn,
  repo_org = intern_gh_org_an(),
  grundsokvag = intern_gh_mapp_an(),
  oppna_rproj = FALSE
)
}
\arguments{
\item{repo_namn}{Reponamnet (utan URL), t.ex. \code{"hamta_data"}.}

\item{repo_org}{Organisation. Standard: \link{rddeploy-config} \code{gh_org}.}

\item{grundsokvag}{Lokal föräldermapp. Standard: \link{rddeploy-config} \code{gh_mapp}.}

\item{oppna_rproj}{Skapa \code{.Rproj}-fil och öppna projektet (kräver \code{usethis}
respektive RStudio).}
}
\value{
Osynligt: sökvägen till det klonade repot.
}
\description{
Klona ett GitHub-repo lokalt
}
