% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_lista_repos}
\alias{github_lista_repos}
\alias{github_lista_repos_analytikernatverket}
\title{Lista repositories hos en GitHub-organisation eller -användare}
\usage{
github_lista_repos(owner = intern_gh_org(), skriv_ut = TRUE)

github_lista_repos_analytikernatverket(
  owner = intern_gh_org_an(),
  skriv_ut = TRUE
)
}
\arguments{
\item{owner}{Organisation/användare. Standard: \link{rddeploy-config} \code{gh_org}.}

\item{skriv_ut}{Skriv reponamnen i konsolen.}
}
\value{
Osynligt (eller synligt om \code{skriv_ut = FALSE}): en tibble med \code{namn},
\code{url}, \code{url_clone}, \code{privat}, \code{arkiverad}.
}
\description{
Lista repositories hos en GitHub-organisation eller -användare
}
