% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_pull_lokalt_repo_fran_github}
\alias{github_pull_lokalt_repo_fran_github}
\alias{github_pull_lokalt_repo_fran_github_analytikernatverket}
\title{Pulla ett eller flera lokala repon från GitHub}
\usage{
github_pull_lokalt_repo_fran_github(repo, grundsokvag = intern_gh_mapp())

github_pull_lokalt_repo_fran_github_analytikernatverket(
  repo,
  grundsokvag = intern_gh_mapp_an()
)
}
\arguments{
\item{repo}{Reponamn, en vektor med flera, eller \code{"*"} för alla mappar i
\code{grundsokvag}.}

\item{grundsokvag}{Föräldermapp. Standard: \link{rddeploy-config} \code{gh_mapp}.}
}
\value{
Osynligt \code{NULL}.
}
\description{
Kör en förkontroll: har repot lokala ändringar hoppas pull över (en pull kan
annars misslyckas tyst).
}
