% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_status_filer_lokalt_repo}
\alias{github_status_filer_lokalt_repo}
\alias{github_status_filer_lokalt_repo_analytikernatverket}
\title{Visa status för ett lokalt klonat repo}
\usage{
github_status_filer_lokalt_repo(repo, grundsokvag = intern_gh_mapp())

github_status_filer_lokalt_repo_analytikernatverket(
  repo,
  grundsokvag = intern_gh_mapp_an()
)
}
\arguments{
\item{repo}{Reponamn.}

\item{grundsokvag}{Föräldermapp där repot ligger. Standard: \link{rddeploy-config}
\code{gh_mapp}.}
}
\value{
Osynligt: \code{gert::git_status()}-tibblen. Skriver en sammanfattning.
}
\description{
Visa status för ett lokalt klonat repo
}
