% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{ppt_lista_rader}
\alias{ppt_lista_rader}
\title{Skriv en kodrad för \code{ppt_lista_fyll_pa()} till urklipp}
\usage{
ppt_lista_rader(ppt_url = "")
}
\arguments{
\item{ppt_url}{Rå-URL till diagramskriptet.}
}
\value{
Osynligt: kodraden som sträng.
}
\description{
Skriv en kodrad för \code{ppt_lista_fyll_pa()} till urklipp
}
