% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rddeploy-package.R
\docType{package}
\name{rddeploy-package}
\alias{rddeploy}
\alias{rddeploy-package}
\title{rddeploy: Deploy-verktyg för Region Dalarnas GitHub, webbrapporter och Shiny-appar}
\description{
Samlar Region Dalarnas arbetsflöden för git och GitHub, publicering av webbrapporter och portaler samt scaffolding och publicering av Shiny-appar. Bygger på gert, gh och gitcreds i stället för handrullade git-anrop. Utbrutet ur func_API.R och func_landningssida_adminportal.R i Region-Dalarna/funktioner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Region-Dalarna/rdpaket}
  \item Report bugs at \url{https://github.com/Region-Dalarna/rdpaket/issues}
}

}
\author{
\strong{Maintainer}: Peter Möller \email{larspetermoller@gmail.com}

Authors:
\itemize{
  \item Peter Möller \email{larspetermoller@gmail.com}
}

Other contributors:
\itemize{
  \item Region Dalarna [copyright holder]
}

}
\keyword{internal}
